Create an experiment in adding "OBJECT" libs
CMake add_library command normally makes static or shared libraries.
Recently a third option was added to make "OBJECT" libraries. These are
not actually build products at all. Instead, when you add_library(foo
OBJECT ...) you get a build target that compiles all of the source to
object files, but then stops short of making an actual shared library
(dll) or archive (.lib). Instead what happens when you compile exe or dll targets against this OBJECT library target, each of the obj files is linked in
directly. This potentially has the advantage of maintaining support for
INCREMENTAL linking which msvc uses to speed up linking when there are
only isolated changes.

It is not clear that this is going to be all that helpful. It is clear
that debug builds default to incremental linking, however this project is
not benefiting because incremental linking does not work through a large
number of static libs.